refactor(cli): Extract output package, remove dead viper wrappers
Extract the structured-data rendering pipeline from internal/cli/output.go
into a new internal/output package with a single public function:
  output.Render(w io.Writer, data any, options output.Options)

Remove dead convenience wrappers (RenderTo, RenderMutationTo), unused
MutationFlags/AddMutationFlags, and 10 unused viper accessor wrappers
(Get, GetBool, GetInt, GetStringSlice, GetStringMap, BindFlagsWithPrefix,
ToolConfigPath, ConfigFileUsed, AllSettings, Debug).

Replace 4 cli.GetString("writ","repo",true) call sites with direct
viper.GetString("writ.repo") calls. Rename internal var 'output' to
'statusOutput' to avoid shadowing the new import.

Closes partially: #238
