+++ internal/cli/viper.go
@@ -119,110 +119,7 @@ func BindFlags(cmd *cobra.Command, toolName string, useSharedConfig bool) error 
 	return bindErr
 }
 
-// BindFlagsWithPrefix binds flags with a custom prefix.
-// Useful for subcommands that need their own config section.
-//
-// Example:
-//
-//	BindFlagsWithPrefix(addCmd.Flags(), "writ.add")
-//	// --overwrite flag → viper key "writ.add.overwrite"
-func BindFlagsWithPrefix(flags *pflag.FlagSet, prefix string) error {
-	var bindErr error
-
-	flags.VisitAll(func(f *pflag.Flag) {
-		if bindErr != nil {
-			return
-		}
-
-		key := prefix + "." + f.Name
-		if err := viper.BindPFlag(key, f); err != nil {
-			bindErr = fmt.Errorf("failed to bind flag %s: %w", f.Name, err)
-		}
-	})
-
-	return bindErr
-}
-
-// Get retrieves a config value with the tool's section prefix.
-// When using shared config, automatically prefixes with tool name.
-//
-// Example:
-//
-//	Get("writ", "repo", true)  → viper.Get("writ.repo")
-//	Get("writ", "repo", false) → viper.Get("repo")
-func Get(toolName, key string, useSharedConfig bool) interface{} {
-	if useSharedConfig {
-		return viper.Get(toolName + "." + key)
-	}
-	return viper.Get(key)
-}
-
-// GetString retrieves a string config value.
-func GetString(toolName, key string, useSharedConfig bool) string {
-	if useSharedConfig {
-		return viper.GetString(toolName + "." + key)
-	}
-	return viper.GetString(key)
-}
-
-// GetBool retrieves a boolean config value.
-func GetBool(toolName, key string, useSharedConfig bool) bool {
-	if useSharedConfig {
-		return viper.GetBool(toolName + "." + key)
-	}
-	return viper.GetBool(key)
-}
-
-// GetInt retrieves an integer config value.
-func GetInt(toolName, key string, useSharedConfig bool) int {
-	if useSharedConfig {
-		return viper.GetInt(toolName + "." + key)
-	}
-	return viper.GetInt(key)
-}
-
-// GetStringSlice retrieves a string slice config value.
-func GetStringSlice(toolName, key string, useSharedConfig bool) []string {
-	if useSharedConfig {
-		return viper.GetStringSlice(toolName + "." + key)
-	}
-	return viper.GetStringSlice(key)
-}
-
-// GetStringMap retrieves a string map config value.
-func GetStringMap(toolName, key string, useSharedConfig bool) map[string]interface{} {
-	if useSharedConfig {
-		return viper.GetStringMap(toolName + "." + key)
-	}
-	return viper.GetStringMap(key)
-}
-
 // SharedConfigPath returns the path to the shared devlore config file.
 func SharedConfigPath() string {
 	return filepath.Join(ConfigHome(), "devlore", "config.yaml")
-}
-
-// ToolConfigPath returns the path to a tool-specific config file.
-func ToolConfigPath(toolName string) string {
-	return filepath.Join(ConfigHome(), toolName, "config.yaml")
-}
-
-// ConfigFileUsed returns the config file path that Viper loaded.
-// Returns empty string if no config file was found.
-func ConfigFileUsed() string {
-	return viper.ConfigFileUsed()
-}
-
-// AllSettings returns all settings as a map.
-func AllSettings() map[string]interface{} {
-	return viper.AllSettings()
-}
-
-// Debug prints current Viper state for debugging.
-func Debug() {
-	Note("Config file: %s", viper.ConfigFileUsed())
-	Note("Settings:")
-	for k, v := range viper.AllSettings() {
-		Note("  %s = %v", k, v)
-	}
 }
